requirements.txt
torch~=2.0.1+cu118
torchvision~=0.15.2+cu118
pillow~=9.4.0
tqdm~=4.65.0
numpy~=1.24.3
visdom
pytorch_msssim